build: Move install-hooks to the end of test script
The hope is that this will prevent spurious build
failures which I'm occassionally running into locally,
but which are not reproducible and which disappear on retry.

+++ tools/git/install-hooks.bash
@@ -3,4 +3,6 @@ set -euo pipefail
 
 DIR=$(realpath "$(dirname "$0")")
 
-cp -v "$DIR/hooks/"* "$DIR/../../.git/hooks/"
+if ! [ -e "$DIR/../../.git/hooks/pre-commit" ]; then
+  cp -v "$DIR/hooks/"* "$DIR/../../.git/hooks/"
+fi
